basicly working evaluator
Need to redo the LispData types not as subclasses of builtins,
they are really throwing mypy for a loop, and just a bit messy in
general.

Probably the thing to do is just implement List as a proper linked list,
Symbol and Bool as their own things. and just use represent Lisp Numbers as Python numbers,
since doing a custom thing would be a lot of work for a not a lot of value.